requirements.txt
# Core requirements for the bot
aiohappyeyeballs>=2.4.4
aiohttp>=3.11.9
discord.py>=2.4.0
PyNaCl>=1.5.0
python-dotenv>=1.0.1
requests>=2.32.3

# Testing requirements
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0


